webapp/requirements.txt
django==1.8.8
pillow==3.1.0
psycopg2==2.6.1
sorl-thumbnail==12.3
django-widget-tweaks==1.4.1
django-ckeditor==4.5.1
django-braces==1.8.1
django-datetime-widget==0.9.3
django-ical==1.3
dj-database-url
pytz
django-bootstrap3==6.2.2
django-dbbackup==2.4.1
django-osm-field==0.2.0
django-easy-select2==1.3.2
django-configurations==1.0
